requirements.txt
certifi==2021.10.8
chardet==3.0.4
idna==2.8
python-dotenv==0.18.0
selenium==3.141.0
typing==3.10.0.0
urllib3==1.24.3
requests~=2.21.0